favoriteRecipe Completed

+++ ITIGraduationProject BackEnd/GraduationProject.BL/DTO/RecipeManger/Read/RatingDTO.cs
@@ -2,13 +2,13 @@
 {
     public class RatingDTO
     {
-        public int Score { get; set; }
+        public decimal Score { get; set; }
     }
 
     public class RatingDto
     {
         public int RatingID { get; set; }
-        public int Score { get; set; }
+        public decimal Score { get; set; }
         public DateTime CreatedAt { get; set; }
         public string UserName { get; set; }
     }
